minor 64bit fix
git-svn-id: https://welly.googlecode.com/svn/trunk@694 ace27a1a-5e51-0410-a197-e9aa1ca61478

+++ WLQuickLookBridge.m
@@ -17,7 +17,7 @@
 - (void)close;
 - (void)makeKeyAndOrderFrontWithEffect:(int)flag canClose:(BOOL)canClose;
 // 10.5 only
-- (void)setURLs:(NSArray *)URLs currentIndex:(unsigned)index preservingDisplayState:(BOOL)flag;
+- (void)setURLs:(NSArray *)URLs currentIndex:(NSUInteger)index preservingDisplayState:(BOOL)flag;
 - (void)setEnableDragNDrop:(BOOL)flag;
 // 10.6 and above
 - (id)sharedPreviewView;
